Explicitly cast from size_t to int32_t

+++ test/oscpp_autocheck.cpp
@@ -364,7 +364,7 @@ public:
     }
 
     Blob(OSCPP::Blob b)
-    : Blob(b.size(), b.data())
+    : Blob(static_cast<int32_t>(b.size()), b.data())
     {}
 
     ~Blob()
